review: fold the call handlers into the core files
No separate CallAutoDiffImplementations.{h,cpp}: the declarations and
external models live with the other dialect-independent handlers in
CoreDialectsAutoDiffImplementations, and the definitions with theirs.
