+++ enzyme/Enzyme/MLIR/Implementations/CMakeLists.txt
@@ -53,7 +53,6 @@ add_public_tablegen_target(EnzymeDerivativesIncGen)
 add_mlir_library(MLIREnzymeImplementations
   AffineAutoDiffOpInterfaceImpl.cpp
   ArithAutoDiffOpInterfaceImpl.cpp
-  CallAutoDiffImplementations.cpp
   CoreDialectsAutoDiffImplementations.cpp
   CoreDialectsAutoDiffRegistration.cpp
   LLVMAutoDiffOpInterfaceImpl.cpp
